dev/requirements.txt
Django==3.2.11
psycopg2-binary==2.9.3
djangorestframework==3.12.4
drf-yasg==1.20.0
python-decouple==3.4
votes==1.0.1